refactor: split waste management ui modules
