+++ packages/plugin-waste-management/src/plugin.translations.de.masterData.entities.ts
@@ -1,0 +1,172 @@
+export const wasteManagementPluginTranslationsDEMasterDataEntities = {
+  regions: {
+    title: 'Regionen',
+    description: 'Pflegen Sie Regionen als zweiten kleinen Stammdatenpfad mit eigenem Dialog.',
+    actions: {
+      openCreate: 'Region anlegen',
+      edit: 'Bearbeiten',
+      cancel: 'Abbrechen',
+      create: 'Region speichern',
+      save: 'Änderungen speichern',
+      saving: 'Speichert…',
+    },
+    fields: { name: 'Name' },
+    dialog: {
+      createTitle: 'Region anlegen',
+      createDescription: 'Erstellen Sie eine neue Region für die geografische Hierarchie.',
+      editTitle: 'Region bearbeiten',
+      editDescription: 'Ändern Sie den Namen der Region.',
+    },
+    messages: {
+      createSuccess: 'Die Waste-Region wurde angelegt.',
+      updateSuccess: 'Die Waste-Region wurde aktualisiert.',
+      saveError: 'Die Waste-Region konnte nicht gespeichert werden.',
+      saveForbidden: 'Für das Speichern von Waste-Regionen fehlt die Berechtigung.',
+    },
+    meta: { regionId: 'Region-ID: {{value}}' },
+  },
+  cities: {
+    title: 'Orte',
+    description: 'Pflegen Sie Orte mit optionaler Regionszuordnung als ersten relationalen Stammdatenpfad.',
+    actions: {
+      openCreate: 'Ort anlegen',
+      edit: 'Bearbeiten',
+      cancel: 'Abbrechen',
+      create: 'Ort speichern',
+      save: 'Änderungen speichern',
+      saving: 'Speichert…',
+    },
+    fields: { name: 'Name', regionId: 'Region', regionUnset: 'Keine Region' },
+    dialog: {
+      createTitle: 'Ort anlegen',
+      createDescription: 'Erstellen Sie einen Ort und ordnen Sie ihn bei Bedarf einer Region zu.',
+      editTitle: 'Ort bearbeiten',
+      editDescription: 'Ändern Sie den Namen oder die Regionszuordnung des Orts.',
+    },
+    messages: {
+      createSuccess: 'Der Waste-Ort wurde angelegt.',
+      updateSuccess: 'Der Waste-Ort wurde aktualisiert.',
+      saveError: 'Der Waste-Ort konnte nicht gespeichert werden.',
+      saveForbidden: 'Für das Speichern von Waste-Orten fehlt die Berechtigung.',
+    },
+    meta: { cityId: 'Ort-ID: {{value}}', regionId: 'Region-ID: {{value}}' },
+  },
+  streets: {
+    title: 'Straßen',
+    description: 'Pflegen Sie Straßen als eigenständigen Adresspfad unterhalb der Orte.',
+    actions: {
+      openCreate: 'Straße anlegen',
+      edit: 'Bearbeiten',
+      cancel: 'Abbrechen',
+      create: 'Straße speichern',
+      save: 'Änderungen speichern',
+      saving: 'Speichert…',
+    },
+    fields: { name: 'Name', cityId: 'Ort', cityUnset: 'Ort auswählen' },
+    dialog: {
+      createTitle: 'Straße anlegen',
+      createDescription: 'Erstellen Sie eine Straße und ordnen Sie sie einem Ort zu.',
+      editTitle: 'Straße bearbeiten',
+      editDescription: 'Ändern Sie den Namen oder den zugeordneten Ort der Straße.',
+    },
+    messages: {
+      createSuccess: 'Die Waste-Straße wurde angelegt.',
+      updateSuccess: 'Die Waste-Straße wurde aktualisiert.',
+      saveError: 'Die Waste-Straße konnte nicht gespeichert werden.',
+      saveForbidden: 'Für das Speichern von Waste-Straßen fehlt die Berechtigung.',
+    },
+    meta: { streetId: 'Straßen-ID: {{value}}', cityId: 'Ort-ID: {{value}}' },
+  },
+  houseNumbers: {
+    title: 'Hausnummern',
+    description: 'Pflegen Sie Hausnummern als letzten direkten Pfad der Adresshierarchie.',
+    actions: {
+      openCreate: 'Hausnummer anlegen',
+      edit: 'Bearbeiten',
+      cancel: 'Abbrechen',
+      create: 'Hausnummer speichern',
+      save: 'Änderungen speichern',
+      saving: 'Speichert…',
+    },
+    fields: { number: 'Hausnummer', streetId: 'Straße', streetUnset: 'Straße auswählen' },
+    dialog: {
+      createTitle: 'Hausnummer anlegen',
+      createDescription: 'Erstellen Sie eine Hausnummer und ordnen Sie sie einer Straße zu.',
+      editTitle: 'Hausnummer bearbeiten',
+      editDescription: 'Ändern Sie Hausnummer oder Straßenbezug.',
+    },
+    messages: {
+      createSuccess: 'Die Waste-Hausnummer wurde angelegt.',
+      updateSuccess: 'Die Waste-Hausnummer wurde aktualisiert.',
+      saveError: 'Die Waste-Hausnummer konnte nicht gespeichert werden.',
+      saveForbidden: 'Für das Speichern von Waste-Hausnummern fehlt die Berechtigung.',
+    },
+    meta: { houseNumberId: 'Hausnummer-ID: {{value}}', streetId: 'Straßen-ID: {{value}}' },
+  },
+  collectionLocations: {
+    title: 'Abholorte',
+    description: 'Pflegen Sie konkrete Abholorte mit geografischer Auswahl bis auf Hausnummernebene.',
+    actions: {
+      openCreate: 'Abholort anlegen',
+      edit: 'Bearbeiten',
+      cancel: 'Abbrechen',
+      create: 'Abholort speichern',
+      save: 'Änderungen speichern',
+      saving: 'Speichert…',
+    },
+    fields: {
+      regionId: 'Region',
+      regionUnset: 'Alle Regionen',
+      cityId: 'Ort',
+      cityUnset: 'Ort auswählen',
+      streetId: 'Straße',
+      streetUnset: 'Alle Straßen',
+      houseNumberId: 'Hausnummer',
+      houseNumberUnset: 'Alle Hausnummern',
+      active: 'Status',
+    },
+    dialog: {
+      createTitle: 'Abholort anlegen',
+      createDescription: 'Erstellen Sie einen neuen Abholort aus Region, Ort, Straße und Hausnummer.',
+      editTitle: 'Abholort bearbeiten',
+      editDescription: 'Ändern Sie geografische Zuordnung und Sichtbarkeit des Abholorts.',
+    },
+    meta: {
+      locationId: 'Abholort-ID: {{value}}',
+      allStreets: 'Alle Straßen',
+      allHouseNumbers: 'Alle Hausnummern',
+    },
+    messages: {
+      createSuccess: 'Der Waste-Abholort wurde angelegt.',
+      updateSuccess: 'Der Waste-Abholort wurde aktualisiert.',
+      saveError: 'Der Waste-Abholort konnte nicht gespeichert werden.',
+      saveForbidden: 'Für das Speichern von Waste-Abholorten fehlt die Berechtigung.',
+    },
+    bulk: {
+      actions: {
+        openAssign: '{{value}} auswählen und Tour zuweisen',
+        selectAllFiltered: 'Alle gefilterten Abholorte auswählen',
+        cancel: 'Abbrechen',
+        assign: 'Tour zuweisen',
+        saving: 'Speichert…',
+      },
+      fields: {
+        tourId: 'Tour',
+        tourUnset: 'Tour auswählen',
+        startDate: 'Startdatum',
+        endDate: 'Enddatum',
+      },
+      dialog: {
+        title: 'Tour gesammelt zuweisen',
+        description: '{{value}} ausgewählte Abholorte werden einer Tour zugeordnet.',
+      },
+      selectedTitle: 'Ausgewählte Abholorte',
+      meta: { selectedCount: '{{value}} ausgewählt' },
+      messages: {
+        assignSuccess: 'Die Waste-Tour-Zuordnungen wurden gesammelt angelegt.',
+        assignError: 'Die Waste-Tour-Zuordnungen konnten nicht gesammelt gespeichert werden.',
+        assignForbidden: 'Für das Sammelspeichern von Waste-Tour-Zuordnungen fehlt die Berechtigung.',
+      },
+    },
+  },
+} as const;
