+++ packages/plugin-waste-management/src/plugin.translations.de.masterData.fractions.ts
@@ -1,0 +1,121 @@
+export const wasteManagementPluginTranslationsDEMasterDataFractions = {
+  title: 'Abfallfraktionen',
+  description: 'Pflegen Sie die sichtbaren Fraktionen als ersten echten Mutationspfad des Plugins.',
+  color: 'Farbe: {{value}}',
+  containerSize: 'Behälter: {{value}}',
+  table: {
+    ariaLabel: 'Tabelle der Waste-Abfallfraktionen',
+    caption: 'Tabelle der Waste-Abfallfraktionen mit Status, Metadaten und Aktionen.',
+    name: 'Fraktion',
+    nameWithContainerSize: 'Name (Gefäßgröße)',
+    selection: 'Auswahl',
+    status: 'Status',
+    color: 'Farbe',
+    containerSize: 'Behälter',
+    translations: 'Übersetzungen',
+    actions: 'Aktionen',
+    noContainerSize: 'Keine Behältergröße',
+    noTranslations: 'Keine Übersetzungen',
+    noDescription: 'Keine Beschreibung',
+    colorSwatch: 'Farbwert {{value}}',
+    selectAllRows: 'Alle Einträge in {{label}} auswählen',
+    selectRow: 'Eintrag {{rowId}} auswählen',
+    selectMobileRow: 'Eintrag {{rowId}} in Kartenansicht auswählen',
+  },
+  actions: {
+    openCreate: 'Fraktion anlegen',
+    edit: 'Bearbeiten',
+    delete: 'Löschen',
+    deleteSelected: 'Markierte löschen',
+    activateStatus: 'Fraktion {{value}} aktivieren',
+    deactivateStatus: 'Fraktion {{value}} deaktivieren',
+    cancel: 'Abbrechen',
+    create: 'Fraktion speichern',
+    save: 'Änderungen speichern',
+    saving: 'Speichert…',
+  },
+  fields: {
+    name: 'Name',
+    translationDe: 'Bezeichnung (DE)',
+    translationEn: 'Bezeichnung (EN)',
+    color: 'Farbe',
+    containerSize: 'Behältergröße',
+    description: 'Beschreibung',
+    active: 'Status',
+  },
+  createView: {
+    title: 'Neue Abfallart anlegen',
+    description: 'Legen Sie eine neue Abfallart an. Die Angaben werden später in Listen, Touren und im Abfallkalender verwendet.',
+    colorPickerLabel: 'Farbe auswählen',
+    sections: {
+      basics: '1. Grunddaten',
+      basicsHint: 'Geben Sie zuerst die wichtigsten Informationen zur neuen Abfallart ein.',
+      presentation: '2. Darstellung',
+      presentationHint: 'Legen Sie fest, wie die Abfallart später erkannt und angezeigt wird.',
+      visibility: '3. Sichtbarkeit',
+      visibilityHint: 'Bestimmen Sie, ob die Abfallart sofort aktiv verwendet werden kann.',
+    },
+    fieldHints: {
+      name: 'So wird die Abfallart später in Listen angezeigt.',
+      description: 'Beschreiben Sie die Abfallart kurz und verständlich. Maximal 300 Zeichen.',
+      translationDe: 'Optional: deutsche Bezeichnung für mehrsprachige Ausgaben.',
+      translationEn: 'Optional: englische Bezeichnung für mehrsprachige Ausgaben.',
+      containerSize: 'Optional: Tragen Sie eine Behältergröße nur ein, wenn diese Abfallart in mehreren Größen geführt wird, zum Beispiel 120 l oder 240 l.',
+      color: 'Diese Farbe wird später zur Wiedererkennung im Kalender und in Tabellen verwendet.',
+      active: 'Nur aktive Abfallarten stehen für die weitere Planung zur Verfügung.',
+    },
+    statusHints: {
+      active: 'Die Abfallart ist sofort für Planung und Anzeige verfügbar.',
+      inactive: 'Die Abfallart bleibt gespeichert, ist aber vorerst nicht aktiv nutzbar.',
+    },
+    validation: {
+      nameRequired: 'Bitte geben Sie einen Namen ein.',
+      colorRequired: 'Bitte geben Sie eine gültige Farbe im Hex-Format an.',
+    },
+    meta: {
+      descriptionCounter: '{{count}} / 300 Zeichen',
+    },
+    actions: {
+      backToList: 'Zurück zur Liste',
+      cancel: 'Abbrechen',
+      savePrimary: 'Abfallart speichern',
+    },
+    preview: {
+      title: 'So sieht die Abfallart später aus',
+      description: 'Die Vorschau aktualisiert sich sofort, während Sie die Angaben eingeben.',
+      placeholderName: 'Neue Abfallart',
+      placeholderDescription: 'Hier erscheint später eine kurze Beschreibung der Abfallart.',
+      placeholderContainerSize: 'Behältergröße noch nicht angegeben',
+      tableTitle: 'Vorschau in der Tabelle',
+      tableBody: 'So erscheint die Abfallart später in der Übersicht.',
+    },
+    help: {
+      title: 'Hinweis',
+      body: 'Sie können alle Angaben später jederzeit wieder ändern.',
+    },
+  },
+  translationBadge: '{{locale}}: {{value}}',
+  dialog: {
+    createTitle: 'Abfallfraktion anlegen',
+    createDescription: 'Erstellen Sie eine neue Fraktion für die weitere Touren- und Kalenderpflege.',
+    editTitle: 'Abfallfraktion bearbeiten',
+    editDescription: 'Ändern Sie Name, Farbe und Sichtbarkeit der Fraktion.',
+  },
+  deleteDialog: {
+    title: 'Abfallfraktion löschen',
+    description: 'Die Abfallfraktion {{value}} wird dauerhaft gelöscht.',
+    confirm: 'Löschen',
+    cancel: 'Abbrechen',
+  },
+  messages: {
+    createSuccess: 'Die Waste-Fraktion wurde angelegt.',
+    updateSuccess: 'Die Waste-Fraktion wurde aktualisiert.',
+    deleteSuccess: 'Die Waste-Fraktion wurde gelöscht.',
+    deletePartialSuccess: '{{count}} von {{total}} Waste-Fraktionen wurden gelöscht.',
+    saveError: 'Die Waste-Fraktion konnte nicht gespeichert werden.',
+    saveForbidden: 'Für das Speichern von Waste-Fraktionen fehlt die Berechtigung.',
+    deleteError: 'Die Waste-Fraktion konnte nicht gelöscht werden.',
+    deleteForbidden: 'Für das Löschen von Waste-Fraktionen fehlt die Berechtigung.',
+    deleteConflict: 'Die Waste-Fraktion kann wegen bestehender Zuordnungen nicht gelöscht werden.',
+  },
+} as const;
