+++ packages/plugin-waste-management/src/plugin.translations.de.masterData.ts
@@ -1,3 +1,5 @@
+import { wasteManagementPluginTranslationsDEMasterDataEntities } from './plugin.translations.de.masterData.entities.js';
+import { wasteManagementPluginTranslationsDEMasterDataFractions } from './plugin.translations.de.masterData.fractions.js';
 import { wasteManagementPluginTranslationsDEMasterDataLocationsWorkspace } from './plugin.translations.de.masterData.locations-workspace.js';
 import {
   createMasterDataTabs,
@@ -22,322 +24,6 @@ export const wasteManagementPluginTranslationsDEMasterData = createWasteManageme
   },
   tabs: createMasterDataTabs('Stammdatenbereiche', 'Fraktionen', 'Abholorte'),
   locationsWorkspace: wasteManagementPluginTranslationsDEMasterDataLocationsWorkspace,
-  fractions: {
-    title: 'Abfallfraktionen',
-    description: 'Pflegen Sie die sichtbaren Fraktionen als ersten echten Mutationspfad des Plugins.',
-    color: 'Farbe: {{value}}',
-    containerSize: 'Behälter: {{value}}',
-    table: {
-      ariaLabel: 'Tabelle der Waste-Abfallfraktionen',
-      caption: 'Tabelle der Waste-Abfallfraktionen mit Status, Metadaten und Aktionen.',
-      name: 'Fraktion',
-      nameWithContainerSize: 'Name (Gefäßgröße)',
-      selection: 'Auswahl',
-      status: 'Status',
-      color: 'Farbe',
-      containerSize: 'Behälter',
-      translations: 'Übersetzungen',
-      actions: 'Aktionen',
-      noContainerSize: 'Keine Behältergröße',
-      noTranslations: 'Keine Übersetzungen',
-      noDescription: 'Keine Beschreibung',
-      colorSwatch: 'Farbwert {{value}}',
-      selectAllRows: 'Alle Einträge in {{label}} auswählen',
-      selectRow: 'Eintrag {{rowId}} auswählen',
-      selectMobileRow: 'Eintrag {{rowId}} in Kartenansicht auswählen',
-    },
-    actions: {
-      openCreate: 'Fraktion anlegen',
-      edit: 'Bearbeiten',
-      delete: 'Löschen',
-      deleteSelected: 'Markierte löschen',
-      activateStatus: 'Fraktion {{value}} aktivieren',
-      deactivateStatus: 'Fraktion {{value}} deaktivieren',
-      cancel: 'Abbrechen',
-      create: 'Fraktion speichern',
-      save: 'Änderungen speichern',
-      saving: 'Speichert…',
-    },
-    fields: {
-      name: 'Name',
-      translationDe: 'Bezeichnung (DE)',
-      translationEn: 'Bezeichnung (EN)',
-      color: 'Farbe',
-      containerSize: 'Behältergröße',
-      description: 'Beschreibung',
-      active: 'Status',
-    },
-    createView: {
-      title: 'Neue Abfallart anlegen',
-      description: 'Legen Sie eine neue Abfallart an. Die Angaben werden später in Listen, Touren und im Abfallkalender verwendet.',
-      colorPickerLabel: 'Farbe auswählen',
-      sections: {
-        basics: '1. Grunddaten',
-        basicsHint: 'Geben Sie zuerst die wichtigsten Informationen zur neuen Abfallart ein.',
-        presentation: '2. Darstellung',
-        presentationHint: 'Legen Sie fest, wie die Abfallart später erkannt und angezeigt wird.',
-        visibility: '3. Sichtbarkeit',
-        visibilityHint: 'Bestimmen Sie, ob die Abfallart sofort aktiv verwendet werden kann.',
-      },
-      fieldHints: {
-        name: 'So wird die Abfallart später in Listen angezeigt.',
-        description: 'Beschreiben Sie die Abfallart kurz und verständlich. Maximal 300 Zeichen.',
-        translationDe: 'Optional: deutsche Bezeichnung für mehrsprachige Ausgaben.',
-        translationEn: 'Optional: englische Bezeichnung für mehrsprachige Ausgaben.',
-        containerSize: 'Optional: Tragen Sie eine Behältergröße nur ein, wenn diese Abfallart in mehreren Größen geführt wird, zum Beispiel 120 l oder 240 l.',
-        color: 'Diese Farbe wird später zur Wiedererkennung im Kalender und in Tabellen verwendet.',
-        active: 'Nur aktive Abfallarten stehen für die weitere Planung zur Verfügung.',
-      },
-      statusHints: {
-        active: 'Die Abfallart ist sofort für Planung und Anzeige verfügbar.',
-        inactive: 'Die Abfallart bleibt gespeichert, ist aber vorerst nicht aktiv nutzbar.',
-      },
-      validation: {
-        nameRequired: 'Bitte geben Sie einen Namen ein.',
-        colorRequired: 'Bitte geben Sie eine gültige Farbe im Hex-Format an.',
-      },
-      meta: {
-        descriptionCounter: '{{count}} / 300 Zeichen',
-      },
-      actions: {
-        backToList: 'Zurück zur Liste',
-        cancel: 'Abbrechen',
-        savePrimary: 'Abfallart speichern',
-      },
-      preview: {
-        title: 'So sieht die Abfallart später aus',
-        description: 'Die Vorschau aktualisiert sich sofort, während Sie die Angaben eingeben.',
-        placeholderName: 'Neue Abfallart',
-        placeholderDescription: 'Hier erscheint später eine kurze Beschreibung der Abfallart.',
-        placeholderContainerSize: 'Behältergröße noch nicht angegeben',
-        tableTitle: 'Vorschau in der Tabelle',
-        tableBody: 'So erscheint die Abfallart später in der Übersicht.',
-      },
-      help: {
-        title: 'Hinweis',
-        body: 'Sie können alle Angaben später jederzeit wieder ändern.',
-      },
-    },
-    translationBadge: '{{locale}}: {{value}}',
-    dialog: {
-      createTitle: 'Abfallfraktion anlegen',
-      createDescription: 'Erstellen Sie eine neue Fraktion für die weitere Touren- und Kalenderpflege.',
-      editTitle: 'Abfallfraktion bearbeiten',
-      editDescription: 'Ändern Sie Name, Farbe und Sichtbarkeit der Fraktion.',
-    },
-    deleteDialog: {
-      title: 'Abfallfraktion löschen',
-      description: 'Die Abfallfraktion {{value}} wird dauerhaft gelöscht.',
-      confirm: 'Löschen',
-      cancel: 'Abbrechen',
-    },
-    messages: {
-      createSuccess: 'Die Waste-Fraktion wurde angelegt.',
-      updateSuccess: 'Die Waste-Fraktion wurde aktualisiert.',
-      deleteSuccess: 'Die Waste-Fraktion wurde gelöscht.',
-      deletePartialSuccess: '{{count}} von {{total}} Waste-Fraktionen wurden gelöscht.',
-      saveError: 'Die Waste-Fraktion konnte nicht gespeichert werden.',
-      saveForbidden: 'Für das Speichern von Waste-Fraktionen fehlt die Berechtigung.',
-      deleteError: 'Die Waste-Fraktion konnte nicht gelöscht werden.',
-      deleteForbidden: 'Für das Löschen von Waste-Fraktionen fehlt die Berechtigung.',
-      deleteConflict: 'Die Waste-Fraktion kann wegen bestehender Zuordnungen nicht gelöscht werden.',
-    },
-  },
-  regions: {
-    title: 'Regionen',
-    description: 'Pflegen Sie Regionen als zweiten kleinen Stammdatenpfad mit eigenem Dialog.',
-    actions: {
-      openCreate: 'Region anlegen',
-      edit: 'Bearbeiten',
-      cancel: 'Abbrechen',
-      create: 'Region speichern',
-      save: 'Änderungen speichern',
-      saving: 'Speichert…',
-    },
-    fields: {
-      name: 'Name',
-    },
-    dialog: {
-      createTitle: 'Region anlegen',
-      createDescription: 'Erstellen Sie eine neue Region für die geografische Hierarchie.',
-      editTitle: 'Region bearbeiten',
-      editDescription: 'Ändern Sie den Namen der Region.',
-    },
-    messages: {
-      createSuccess: 'Die Waste-Region wurde angelegt.',
-      updateSuccess: 'Die Waste-Region wurde aktualisiert.',
-      saveError: 'Die Waste-Region konnte nicht gespeichert werden.',
-      saveForbidden: 'Für das Speichern von Waste-Regionen fehlt die Berechtigung.',
-    },
-    meta: {
-      regionId: 'Region-ID: {{value}}',
-    },
-  },
-  cities: {
-    title: 'Orte',
-    description: 'Pflegen Sie Orte mit optionaler Regionszuordnung als ersten relationalen Stammdatenpfad.',
-    actions: {
-      openCreate: 'Ort anlegen',
-      edit: 'Bearbeiten',
-      cancel: 'Abbrechen',
-      create: 'Ort speichern',
-      save: 'Änderungen speichern',
-      saving: 'Speichert…',
-    },
-    fields: {
-      name: 'Name',
-      regionId: 'Region',
-      regionUnset: 'Keine Region',
-    },
-    dialog: {
-      createTitle: 'Ort anlegen',
-      createDescription: 'Erstellen Sie einen Ort und ordnen Sie ihn bei Bedarf einer Region zu.',
-      editTitle: 'Ort bearbeiten',
-      editDescription: 'Ändern Sie den Namen oder die Regionszuordnung des Orts.',
-    },
-    messages: {
-      createSuccess: 'Der Waste-Ort wurde angelegt.',
-      updateSuccess: 'Der Waste-Ort wurde aktualisiert.',
-      saveError: 'Der Waste-Ort konnte nicht gespeichert werden.',
-      saveForbidden: 'Für das Speichern von Waste-Orten fehlt die Berechtigung.',
-    },
-    meta: {
-      cityId: 'Ort-ID: {{value}}',
-      regionId: 'Region-ID: {{value}}',
-    },
-  },
-  streets: {
-    title: 'Straßen',
-    description: 'Pflegen Sie Straßen als eigenständigen Adresspfad unterhalb der Orte.',
-    actions: {
-      openCreate: 'Straße anlegen',
-      edit: 'Bearbeiten',
-      cancel: 'Abbrechen',
-      create: 'Straße speichern',
-      save: 'Änderungen speichern',
-      saving: 'Speichert…',
-    },
-    fields: {
-      name: 'Name',
-      cityId: 'Ort',
-      cityUnset: 'Ort auswählen',
-    },
-    dialog: {
-      createTitle: 'Straße anlegen',
-      createDescription: 'Erstellen Sie eine Straße und ordnen Sie sie einem Ort zu.',
-      editTitle: 'Straße bearbeiten',
-      editDescription: 'Ändern Sie den Namen oder den zugeordneten Ort der Straße.',
-    },
-    messages: {
-      createSuccess: 'Die Waste-Straße wurde angelegt.',
-      updateSuccess: 'Die Waste-Straße wurde aktualisiert.',
-      saveError: 'Die Waste-Straße konnte nicht gespeichert werden.',
-      saveForbidden: 'Für das Speichern von Waste-Straßen fehlt die Berechtigung.',
-    },
-    meta: {
-      streetId: 'Straßen-ID: {{value}}',
-      cityId: 'Ort-ID: {{value}}',
-    },
-  },
-  houseNumbers: {
-    title: 'Hausnummern',
-    description: 'Pflegen Sie Hausnummern als letzten direkten Pfad der Adresshierarchie.',
-    actions: {
-      openCreate: 'Hausnummer anlegen',
-      edit: 'Bearbeiten',
-      cancel: 'Abbrechen',
-      create: 'Hausnummer speichern',
-      save: 'Änderungen speichern',
-      saving: 'Speichert…',
-    },
-    fields: {
-      number: 'Hausnummer',
-      streetId: 'Straße',
-      streetUnset: 'Straße auswählen',
-    },
-    dialog: {
-      createTitle: 'Hausnummer anlegen',
-      createDescription: 'Erstellen Sie eine Hausnummer und ordnen Sie sie einer Straße zu.',
-      editTitle: 'Hausnummer bearbeiten',
-      editDescription: 'Ändern Sie Hausnummer oder Straßenbezug.',
-    },
-    messages: {
-      createSuccess: 'Die Waste-Hausnummer wurde angelegt.',
-      updateSuccess: 'Die Waste-Hausnummer wurde aktualisiert.',
-      saveError: 'Die Waste-Hausnummer konnte nicht gespeichert werden.',
-      saveForbidden: 'Für das Speichern von Waste-Hausnummern fehlt die Berechtigung.',
-    },
-    meta: {
-      houseNumberId: 'Hausnummer-ID: {{value}}',
-      streetId: 'Straßen-ID: {{value}}',
-    },
-  },
-  collectionLocations: {
-    title: 'Abholorte',
-    description: 'Pflegen Sie konkrete Abholorte mit geografischer Auswahl bis auf Hausnummernebene.',
-    actions: {
-      openCreate: 'Abholort anlegen',
-      edit: 'Bearbeiten',
-      cancel: 'Abbrechen',
-      create: 'Abholort speichern',
-      save: 'Änderungen speichern',
-      saving: 'Speichert…',
-    },
-    fields: {
-      regionId: 'Region',
-      regionUnset: 'Alle Regionen',
-      cityId: 'Ort',
-      cityUnset: 'Ort auswählen',
-      streetId: 'Straße',
-      streetUnset: 'Alle Straßen',
-      houseNumberId: 'Hausnummer',
-      houseNumberUnset: 'Alle Hausnummern',
-      active: 'Status',
-    },
-    dialog: {
-      createTitle: 'Abholort anlegen',
-      createDescription: 'Erstellen Sie einen neuen Abholort aus Region, Ort, Straße und Hausnummer.',
-      editTitle: 'Abholort bearbeiten',
-      editDescription: 'Ändern Sie geografische Zuordnung und Sichtbarkeit des Abholorts.',
-    },
-    meta: {
-      locationId: 'Abholort-ID: {{value}}',
-      allStreets: 'Alle Straßen',
-      allHouseNumbers: 'Alle Hausnummern',
-    },
-    messages: {
-      createSuccess: 'Der Waste-Abholort wurde angelegt.',
-      updateSuccess: 'Der Waste-Abholort wurde aktualisiert.',
-      saveError: 'Der Waste-Abholort konnte nicht gespeichert werden.',
-      saveForbidden: 'Für das Speichern von Waste-Abholorten fehlt die Berechtigung.',
-    },
-    bulk: {
-      actions: {
-        openAssign: '{{value}} auswählen und Tour zuweisen',
-        selectAllFiltered: 'Alle gefilterten Abholorte auswählen',
-        cancel: 'Abbrechen',
-        assign: 'Tour zuweisen',
-        saving: 'Speichert…',
-      },
-      fields: {
-        tourId: 'Tour',
-        tourUnset: 'Tour auswählen',
-        startDate: 'Startdatum',
-        endDate: 'Enddatum',
-      },
-      dialog: {
-        title: 'Tour gesammelt zuweisen',
-        description: '{{value}} ausgewählte Abholorte werden einer Tour zugeordnet.',
-      },
-      selectedTitle: 'Ausgewählte Abholorte',
-      meta: {
-        selectedCount: '{{value}} ausgewählt',
-      },
-      messages: {
-        assignSuccess: 'Die Waste-Tour-Zuordnungen wurden gesammelt angelegt.',
-        assignError: 'Die Waste-Tour-Zuordnungen konnten nicht gesammelt gespeichert werden.',
-        assignForbidden: 'Für das Sammelspeichern von Waste-Tour-Zuordnungen fehlt die Berechtigung.',
-      },
-    },
-  },
+  fractions: wasteManagementPluginTranslationsDEMasterDataFractions,
+  ...wasteManagementPluginTranslationsDEMasterDataEntities,
 });
