+++ packages/plugin-waste-management/src/plugin.translations.en.masterData.fractions.ts
@@ -1,0 +1,121 @@
+export const wasteManagementPluginTranslationsENMasterDataFractions = {
+  title: 'Waste fractions',
+  description: 'Maintain visible fractions as the first real mutation path of the plugin.',
+  color: 'Color: {{value}}',
+  containerSize: 'Container: {{value}}',
+  table: {
+    ariaLabel: 'Table of waste fractions',
+    caption: 'Table of waste fractions with status, metadata, and actions.',
+    name: 'Fraction',
+    nameWithContainerSize: 'Name (Container size)',
+    selection: 'Selection',
+    status: 'Status',
+    color: 'Color',
+    containerSize: 'Container',
+    translations: 'Translations',
+    actions: 'Actions',
+    noContainerSize: 'No container size',
+    noTranslations: 'No translations',
+    noDescription: 'No description',
+    colorSwatch: 'Color value {{value}}',
+    selectAllRows: 'Select all entries in {{label}}',
+    selectRow: 'Select entry {{rowId}}',
+    selectMobileRow: 'Select entry {{rowId}} in card view',
+  },
+  actions: {
+    openCreate: 'Create fraction',
+    edit: 'Edit',
+    delete: 'Delete',
+    deleteSelected: 'Delete selected',
+    activateStatus: 'Activate fraction {{value}}',
+    deactivateStatus: 'Deactivate fraction {{value}}',
+    cancel: 'Cancel',
+    create: 'Save fraction',
+    save: 'Save changes',
+    saving: 'Saving…',
+  },
+  fields: {
+    name: 'Name',
+    translationDe: 'Label (DE)',
+    translationEn: 'Label (EN)',
+    color: 'Color',
+    containerSize: 'Container size',
+    description: 'Description',
+    active: 'Status',
+  },
+  createView: {
+    title: 'Create new waste type',
+    description: 'Create a new waste type. The details will later be used in lists, tours, and the waste calendar.',
+    colorPickerLabel: 'Choose color',
+    sections: {
+      basics: '1. Basic details',
+      basicsHint: 'Start with the most important information about the new waste type.',
+      presentation: '2. Presentation',
+      presentationHint: 'Define how the waste type should be recognized and displayed later.',
+      visibility: '3. Visibility',
+      visibilityHint: 'Decide whether the waste type can be used right away.',
+    },
+    fieldHints: {
+      name: 'This is how the waste type will later appear in lists.',
+      description: 'Describe the waste type briefly and clearly. Maximum 300 characters.',
+      translationDe: 'Optional: German label for multilingual output.',
+      translationEn: 'Optional: English label for multilingual output.',
+      containerSize: 'Optional: Only enter a container size if this waste type exists in multiple sizes, for example 120 l or 240 l.',
+      color: 'This color will later help users recognize the waste type in calendars and tables.',
+      active: 'Only active waste types are available for further planning.',
+    },
+    statusHints: {
+      active: 'The waste type is immediately available for planning and display.',
+      inactive: 'The waste type stays saved, but cannot be actively used yet.',
+    },
+    validation: {
+      nameRequired: 'Please enter a name.',
+      colorRequired: 'Please enter a valid hex color.',
+    },
+    meta: {
+      descriptionCounter: '{{count}} / 300 characters',
+    },
+    actions: {
+      backToList: 'Back to list',
+      cancel: 'Cancel',
+      savePrimary: 'Save waste type',
+    },
+    preview: {
+      title: 'This is how the waste type will look later',
+      description: 'The preview updates immediately while you enter the details.',
+      placeholderName: 'New waste type',
+      placeholderDescription: 'A short description of the waste type will appear here later.',
+      placeholderContainerSize: 'Container size not set yet',
+      tableTitle: 'Table preview',
+      tableBody: 'This is how the waste type will later appear in the overview.',
+    },
+    help: {
+      title: 'Note',
+      body: 'You can change all details again later at any time.',
+    },
+  },
+  translationBadge: '{{locale}}: {{value}}',
+  dialog: {
+    createTitle: 'Create waste fraction',
+    createDescription: 'Create a new fraction for further tour and calendar maintenance.',
+    editTitle: 'Edit waste fraction',
+    editDescription: 'Change name, color and visibility of the fraction.',
+  },
+  deleteDialog: {
+    title: 'Delete waste fraction',
+    description: 'The waste fraction {{value}} will be deleted permanently.',
+    confirm: 'Delete',
+    cancel: 'Cancel',
+  },
+  messages: {
+    createSuccess: 'The waste fraction was created.',
+    updateSuccess: 'The waste fraction was updated.',
+    deleteSuccess: 'The waste fraction was deleted.',
+    deletePartialSuccess: '{{count}} of {{total}} waste fractions were deleted.',
+    saveError: 'The waste fraction could not be saved.',
+    saveForbidden: 'Missing permission to save waste fractions.',
+    deleteError: 'The waste fraction could not be deleted.',
+    deleteForbidden: 'Missing permission to delete waste fractions.',
+    deleteConflict: 'The waste fraction cannot be deleted because assignments still exist.',
+  },
+} as const;
