+++ packages/plugin-waste-management/src/plugin.translations.en.masterData.ts
@@ -1,3 +1,5 @@
+import { wasteManagementPluginTranslationsENMasterDataEntities } from './plugin.translations.en.masterData.entities.js';
+import { wasteManagementPluginTranslationsENMasterDataFractions } from './plugin.translations.en.masterData.fractions.js';
 import { wasteManagementPluginTranslationsENMasterDataLocationsWorkspace } from './plugin.translations.en.masterData.locations-workspace.js';
 import {
   createMasterDataTabs,
@@ -22,322 +24,6 @@ export const wasteManagementPluginTranslationsENMasterData = createWasteManageme
   },
   tabs: createMasterDataTabs('Master-data areas', 'Fractions', 'Locations'),
   locationsWorkspace: wasteManagementPluginTranslationsENMasterDataLocationsWorkspace,
-  fractions: {
-    title: 'Waste fractions',
-    description: 'Maintain visible fractions as the first real mutation path of the plugin.',
-    color: 'Color: {{value}}',
-    containerSize: 'Container: {{value}}',
-    table: {
-      ariaLabel: 'Table of waste fractions',
-      caption: 'Table of waste fractions with status, metadata, and actions.',
-      name: 'Fraction',
-      nameWithContainerSize: 'Name (Container size)',
-      selection: 'Selection',
-      status: 'Status',
-      color: 'Color',
-      containerSize: 'Container',
-      translations: 'Translations',
-      actions: 'Actions',
-      noContainerSize: 'No container size',
-      noTranslations: 'No translations',
-      noDescription: 'No description',
-      colorSwatch: 'Color value {{value}}',
-      selectAllRows: 'Select all entries in {{label}}',
-      selectRow: 'Select entry {{rowId}}',
-      selectMobileRow: 'Select entry {{rowId}} in card view',
-    },
-    actions: {
-      openCreate: 'Create fraction',
-      edit: 'Edit',
-      delete: 'Delete',
-      deleteSelected: 'Delete selected',
-      activateStatus: 'Activate fraction {{value}}',
-      deactivateStatus: 'Deactivate fraction {{value}}',
-      cancel: 'Cancel',
-      create: 'Save fraction',
-      save: 'Save changes',
-      saving: 'Saving…',
-    },
-    fields: {
-      name: 'Name',
-      translationDe: 'Label (DE)',
-      translationEn: 'Label (EN)',
-      color: 'Color',
-      containerSize: 'Container size',
-      description: 'Description',
-      active: 'Status',
-    },
-    createView: {
-      title: 'Create new waste type',
-      description: 'Create a new waste type. The details will later be used in lists, tours, and the waste calendar.',
-      colorPickerLabel: 'Choose color',
-      sections: {
-        basics: '1. Basic details',
-        basicsHint: 'Start with the most important information about the new waste type.',
-        presentation: '2. Presentation',
-        presentationHint: 'Define how the waste type should be recognized and displayed later.',
-        visibility: '3. Visibility',
-        visibilityHint: 'Decide whether the waste type can be used right away.',
-      },
-      fieldHints: {
-        name: 'This is how the waste type will later appear in lists.',
-        description: 'Describe the waste type briefly and clearly. Maximum 300 characters.',
-        translationDe: 'Optional: German label for multilingual output.',
-        translationEn: 'Optional: English label for multilingual output.',
-        containerSize: 'Optional: Only enter a container size if this waste type exists in multiple sizes, for example 120 l or 240 l.',
-        color: 'This color will later help users recognize the waste type in calendars and tables.',
-        active: 'Only active waste types are available for further planning.',
-      },
-      statusHints: {
-        active: 'The waste type is immediately available for planning and display.',
-        inactive: 'The waste type stays saved, but cannot be actively used yet.',
-      },
-      validation: {
-        nameRequired: 'Please enter a name.',
-        colorRequired: 'Please enter a valid hex color.',
-      },
-      meta: {
-        descriptionCounter: '{{count}} / 300 characters',
-      },
-      actions: {
-        backToList: 'Back to list',
-        cancel: 'Cancel',
-        savePrimary: 'Save waste type',
-      },
-      preview: {
-        title: 'This is how the waste type will look later',
-        description: 'The preview updates immediately while you enter the details.',
-        placeholderName: 'New waste type',
-        placeholderDescription: 'A short description of the waste type will appear here later.',
-        placeholderContainerSize: 'Container size not set yet',
-        tableTitle: 'Table preview',
-        tableBody: 'This is how the waste type will later appear in the overview.',
-      },
-      help: {
-        title: 'Note',
-        body: 'You can change all details again later at any time.',
-      },
-    },
-    translationBadge: '{{locale}}: {{value}}',
-    dialog: {
-      createTitle: 'Create waste fraction',
-      createDescription: 'Create a new fraction for further tour and calendar maintenance.',
-      editTitle: 'Edit waste fraction',
-      editDescription: 'Change name, color and visibility of the fraction.',
-    },
-    deleteDialog: {
-      title: 'Delete waste fraction',
-      description: 'The waste fraction {{value}} will be deleted permanently.',
-      confirm: 'Delete',
-      cancel: 'Cancel',
-    },
-    messages: {
-      createSuccess: 'The waste fraction was created.',
-      updateSuccess: 'The waste fraction was updated.',
-      deleteSuccess: 'The waste fraction was deleted.',
-      deletePartialSuccess: '{{count}} of {{total}} waste fractions were deleted.',
-      saveError: 'The waste fraction could not be saved.',
-      saveForbidden: 'Missing permission to save waste fractions.',
-      deleteError: 'The waste fraction could not be deleted.',
-      deleteForbidden: 'Missing permission to delete waste fractions.',
-      deleteConflict: 'The waste fraction cannot be deleted because assignments still exist.',
-    },
-  },
-  regions: {
-    title: 'Regions',
-    description: 'Maintain regions as the second small master-data path with its own dialog.',
-    actions: {
-      openCreate: 'Create region',
-      edit: 'Edit',
-      cancel: 'Cancel',
-      create: 'Save region',
-      save: 'Save changes',
-      saving: 'Saving…',
-    },
-    fields: {
-      name: 'Name',
-    },
-    dialog: {
-      createTitle: 'Create region',
-      createDescription: 'Create a new region for the geographic hierarchy.',
-      editTitle: 'Edit region',
-      editDescription: 'Change the region name.',
-    },
-    messages: {
-      createSuccess: 'The waste region was created.',
-      updateSuccess: 'The waste region was updated.',
-      saveError: 'The waste region could not be saved.',
-      saveForbidden: 'Missing permission to save waste regions.',
-    },
-    meta: {
-      regionId: 'Region ID: {{value}}',
-    },
-  },
-  cities: {
-    title: 'Cities',
-    description: 'Maintain cities with optional region assignment as the first relational master-data path.',
-    actions: {
-      openCreate: 'Create city',
-      edit: 'Edit',
-      cancel: 'Cancel',
-      create: 'Save city',
-      save: 'Save changes',
-      saving: 'Saving…',
-    },
-    fields: {
-      name: 'Name',
-      regionId: 'Region',
-      regionUnset: 'No region',
-    },
-    dialog: {
-      createTitle: 'Create city',
-      createDescription: 'Create a city and assign it to a region when needed.',
-      editTitle: 'Edit city',
-      editDescription: 'Change the city name or region assignment.',
-    },
-    messages: {
-      createSuccess: 'The waste city was created.',
-      updateSuccess: 'The waste city was updated.',
-      saveError: 'The waste city could not be saved.',
-      saveForbidden: 'Missing permission to save waste cities.',
-    },
-    meta: {
-      cityId: 'City ID: {{value}}',
-      regionId: 'Region ID: {{value}}',
-    },
-  },
-  streets: {
-    title: 'Streets',
-    description: 'Maintain streets as a dedicated address path below cities.',
-    actions: {
-      openCreate: 'Create street',
-      edit: 'Edit',
-      cancel: 'Cancel',
-      create: 'Save street',
-      save: 'Save changes',
-      saving: 'Saving…',
-    },
-    fields: {
-      name: 'Name',
-      cityId: 'City',
-      cityUnset: 'Select city',
-    },
-    dialog: {
-      createTitle: 'Create street',
-      createDescription: 'Create a street and assign it to a city.',
-      editTitle: 'Edit street',
-      editDescription: 'Adjust the street name or its linked city.',
-    },
-    messages: {
-      createSuccess: 'The waste street was created.',
-      updateSuccess: 'The waste street was updated.',
-      saveError: 'The waste street could not be saved.',
-      saveForbidden: 'Missing permission to save waste streets.',
-    },
-    meta: {
-      streetId: 'Street ID: {{value}}',
-      cityId: 'City ID: {{value}}',
-    },
-  },
-  houseNumbers: {
-    title: 'House numbers',
-    description: 'Maintain house numbers as the final direct path of the address hierarchy.',
-    actions: {
-      openCreate: 'Create house number',
-      edit: 'Edit',
-      cancel: 'Cancel',
-      create: 'Save house number',
-      save: 'Save changes',
-      saving: 'Saving…',
-    },
-    fields: {
-      number: 'House number',
-      streetId: 'Street',
-      streetUnset: 'Select street',
-    },
-    dialog: {
-      createTitle: 'Create house number',
-      createDescription: 'Create a house number and assign it to a street.',
-      editTitle: 'Edit house number',
-      editDescription: 'Adjust the house number or its linked street.',
-    },
-    messages: {
-      createSuccess: 'The waste house number was created.',
-      updateSuccess: 'The waste house number was updated.',
-      saveError: 'The waste house number could not be saved.',
-      saveForbidden: 'Missing permission to save waste house numbers.',
-    },
-    meta: {
-      houseNumberId: 'House number ID: {{value}}',
-      streetId: 'Street ID: {{value}}',
-    },
-  },
-  collectionLocations: {
-    title: 'Collection locations',
-    description: 'Maintain concrete collection locations with geographic selection down to house-number level.',
-    actions: {
-      openCreate: 'Create collection location',
-      edit: 'Edit',
-      cancel: 'Cancel',
-      create: 'Save collection location',
-      save: 'Save changes',
-      saving: 'Saving…',
-    },
-    fields: {
-      regionId: 'Region',
-      regionUnset: 'All regions',
-      cityId: 'City',
-      cityUnset: 'Select city',
-      streetId: 'Street',
-      streetUnset: 'All streets',
-      houseNumberId: 'House number',
-      houseNumberUnset: 'All house numbers',
-      active: 'Status',
-    },
-    dialog: {
-      createTitle: 'Create collection location',
-      createDescription: 'Create a new collection location from region, city, street and house number.',
-      editTitle: 'Edit collection location',
-      editDescription: 'Adjust the geographic assignment and visibility of the collection location.',
-    },
-    meta: {
-      locationId: 'Location ID: {{value}}',
-      allStreets: 'All streets',
-      allHouseNumbers: 'All house numbers',
-    },
-    messages: {
-      createSuccess: 'The waste collection location was created.',
-      updateSuccess: 'The waste collection location was updated.',
-      saveError: 'The waste collection location could not be saved.',
-      saveForbidden: 'Missing permission to save waste collection locations.',
-    },
-    bulk: {
-      actions: {
-        openAssign: 'Assign tour to {{value}} selected',
-        selectAllFiltered: 'Select all filtered collection locations',
-        cancel: 'Cancel',
-        assign: 'Assign tour',
-        saving: 'Saving…',
-      },
-      fields: {
-        tourId: 'Tour',
-        tourUnset: 'Select tour',
-        startDate: 'Start date',
-        endDate: 'End date',
-      },
-      dialog: {
-        title: 'Bulk assign tour',
-        description: 'Assign a tour to {{value}} selected collection locations.',
-      },
-      selectedTitle: 'Selected collection locations',
-      meta: {
-        selectedCount: '{{value}} selected',
-      },
-      messages: {
-        assignSuccess: 'The waste tour assignments were created in bulk.',
-        assignError: 'The waste tour assignments could not be created in bulk.',
-        assignForbidden: 'Missing permission to create waste tour assignments in bulk.',
-      },
-    },
-  },
+  fractions: wasteManagementPluginTranslationsENMasterDataFractions,
+  ...wasteManagementPluginTranslationsENMasterDataEntities,
 });
