+++ packages/plugin-waste-management/src/waste-management.tours.content.parts.tsx
@@ -3,6 +3,34 @@ import type { WasteTourRecord } from '@sva/plugin-sdk';
 import { usePluginTranslation } from '@sva/plugin-sdk';
 import { StudioConfirmDialog } from '@sva/studio-ui-react';
 
+import type { WasteManagementMasterDataOverview, WasteManagementSchedulingOverview } from './waste-management.api.js';
+
+export type WasteToursContentProps = {
+  readonly assignmentContextLoading: boolean;
+  readonly message: import('./waste-management.page.support.js').StatusMessage | null;
+  readonly tours: readonly WasteTourRecord[];
+  readonly fractions: readonly { readonly id: string; readonly name: string }[];
+  readonly masterDataOverview: WasteManagementMasterDataOverview | null;
+  readonly schedulingOverview: WasteManagementSchedulingOverview | null;
+  readonly onOpenCreateDialog: () => void;
+  readonly onOpenEditDialog: (tour: WasteTourRecord) => void;
+  readonly onOpenCreateAssignmentsDialog: (tour: WasteTourRecord) => void;
+  readonly onOpenEditAssignmentsDialog: (tour: WasteTourRecord, linkId: string) => void;
+  readonly onOpenCalendar: (tour: WasteTourRecord) => void;
+  readonly onToggleTourStatus: (tour: WasteTourRecord, nextActive: boolean) => Promise<void>;
+  readonly onDeleteTour: (tour: WasteTourRecord) => Promise<void>;
+  readonly onDeleteTours: (tourIds: readonly string[]) => Promise<void>;
+  readonly saving?: boolean;
+  readonly page: number;
+  readonly pageSize: number;
+  readonly query: string;
+  readonly status: 'all' | 'active' | 'inactive';
+  readonly onPageChange: (page: number) => void;
+  readonly onPageSizeChange: (pageSize: number) => void;
+  readonly onQueryChange: (value: string) => void;
+  readonly onStatusChange: (value: 'all' | 'active' | 'inactive') => void;
+};
+
 type UseWasteToursSelectionStateArgs = {
   readonly tours: readonly WasteTourRecord[];
   readonly page: number;
